Remove TenantKey type
The goal was to represent the TenantKey as a value object because it has specific rules that can be encapsulated. The concern was that it felt out of place with the rest of the API where config keys and sections are named using plain strings. Removing the TenantKey type feels like the right thing to do to keep  consistency and ease of use.